# [perf] measure paged attention token prefill/generate time