--- a/MaxText/tests/inference/test_llama2_7b_bf16.sh
+++ b/MaxText/tests/inference/test_llama2_7b_bf16.sh
@@ -1,6 +1,5 @@
 #!/bin/bash
 
-# Define the arguments in an array
 args=(
   "MaxText/decode.py"
   "MaxText/configs/base.yml"
@@ -22,6 +21,5 @@ args=(
   "pagedattn_pages_per_compute_block=4"
 )
 
-# Execute the Python script with the arguments
 python "${args[@]}"