```
$ ./MaxText/tests/inference/test_llama2_7b_int8.sh
```
```
prefill: stream=0, #tokens=4/16, time=7906.16 ms
generate: token=#17, time=8804.44 ms
generate: token=#18, time=2.51 ms
generate: token=#19, time=1.99 ms
generate: token=#20, time=1.74 ms
generate: token=#21, time=1.99 ms
generate: token=#22, time=2.20 ms
generate: token=#23, time=1.74 ms
generate: token=#24, time=1.74 ms
generate: token=#25, time=1.77 ms
generate: token=#26, time=3.11 ms
generate: token=#27, time=11.25 ms
generate: token=#28, time=11.09 ms
generate: token=#29, time=11.12 ms
generate: token=#30, time=11.07 ms
generate: token=#31, time=11.08 ms
generate: token=#32, time=11.13 ms
```

```
$ ./MaxText/tests/inference/test_llama2_7b_bf16.sh
```
```
prefill: stream=0, #tokens=4/16, time=3002.58 ms
generate: token=#17, time=3368.66 ms
generate: token=#18, time=1021.62 ms
generate: token=#19, time=1.64 ms
generate: token=#20, time=1.67 ms
generate: token=#21, time=1.37 ms
generate: token=#22, time=1.76 ms
generate: token=#23, time=1.42 ms
generate: token=#24, time=1.43 ms
generate: token=#25, time=1.37 ms
generate: token=#26, time=1.41 ms
generate: token=#27, time=4.33 ms
generate: token=#28, time=9.34 ms
generate: token=#29, time=9.35 ms
generate: token=#30, time=9.27 ms
generate: token=#31, time=9.29 ms
generate: token=#32, time=9.32 ms

```

# Description

Start with a short description of what the PR does and how this is a change from
the past.

The rest of the description includes relevant details and context, examples:
- why is this change being made,
- the problem being solved and any relevant context,
- why this is a good solution,
- some information about the specific implementation,
- shortcomings of the solution and possible future improvements.

If the change fixes a bug or a Github issue, please include a link, e.g.,:
FIXES: b/123456
FIXES: #123456

*Notice 1:* Once all tests pass, the "pull ready" label will automatically be assigned.
This label is used for administrative purposes. Please do not add it manually.

*Notice 2:* For external contributions, our settings currently require an approval from a MaxText maintainer to trigger CI tests.

# Tests

Please describe how you tested this change, and include any instructions and/or
commands to reproduce.

# Checklist

Before submitting this PR, please make sure (put X in square brackets):
- [x] I have performed a self-review of my code.
- [x] I have necessary comments in my code, particularly in hard-to-understand areas.
- [x] I have run end-to-end tests tests and provided workload links above if applicable.
- [x] I have made or will make corresponding changes to the doc if needed.